SMTP: add notification handler, ProtocolError etc
